+++ src/spy_face.h
@@ -9,10 +9,14 @@ typedef struct {
   int second;
   int day;
   int month;
+  int year;
   int weekday;
+  int date_format;
   int battery_percent;
   bool bluetooth_connected;
   bool backlight_on;
+  bool quiet_time_active;
+  bool analogue_seconds_enabled;
   bool twenty_four_hour_style;
 } SpyFaceState;
 
@@ -20,3 +24,4 @@ Layer *spy_face_layer_create(GRect frame);
 void spy_face_layer_destroy(Layer *layer);
 void spy_face_layer_set_state(Layer *layer, const SpyFaceState *state);
 void spy_face_layer_toggle_mode(Layer *layer);
+bool spy_face_layer_wants_second_ticks(Layer *layer);